requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
cloudpathlib==0.16.0
cmake==3.29.0.1
comm==0.2.2
confection==0.1.4
ConfigArgParse==1.7
contexttimer==0.3.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
decord==0.6.0
distlib==0.3.8
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.4
fastjsonschema==2.20.0
fightingcv-attention==1.0.0
filelock==3.13.3
Flask==3.0.3
fonttools==4.50.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
huggingface-hub==0.22.2
identify==2.5.35
idna==3.6
imageio==2.34.0
importlib_metadata==8.0.0
iopath==0.1.10
ipython==8.23.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kaggle==1.6.8
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lightning-utilities==0.11.2
lit==18.1.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.8.0
numpy==1.23.0
omegaconf==2.3.0
open3d==0.18.0
opencv-fixer==0.2.5
opencv-python==4.9.0.80
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
portalocker==2.8.2
pre-commit==3.7.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pyglet==2.0.14
Pygments==2.17.2
PyOpenGL==3.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
salesforce-lavis==1.0.2
scikit-image==0.22.0
scikit-learn==1.5.1
scipy==1.12.0
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
smplx==0.1.28
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2024.2.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torchmetrics==1.3.2
torchsummary==1.5.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.26.1
trimesh==4.2.4
triton==2.0.0
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.11
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
